// packages/@aws-cdk/integ-runner/lib/workers/common.d.ts
export type AssertionResults = {
    [id: string]: AssertionResult;
};
/**
 * The result of an individual assertion
 */
export interface AssertionResult {
    /**
     * The assertion message. If the assertion failed, this will
     * include the reason.
     */
    readonly message: string;
    /**
     * Whether the assertion succeeded or failed
     */
    readonly status: 'success' | 'fail';
}
/**
 * Config for an integration test
 */
export interface IntegTestWorkerConfig extends IntegTestInfo {
    /**
     * A list of any destructive changes
     *
     * @default []
     */
    readonly destructiveChanges?: DestructiveChange[];
}
/**
 * Information on any destructive changes
 */
export interface DestructiveChange {
    /**
     * The logicalId of the resource with a destructive change
     */
    readonly logicalId: string;
    /**
     * The name of the stack that contains the destructive change
     */
    readonly stackName: string;
    /**
     * The impact of the destructive change
     */
    readonly impact: ResourceImpact;
}
// ...
export declare enum DiagnosticReason {
    /**
     * The integration test failed because there
     * is not existing snapshot
     */
    NO_SNAPSHOT = "NO_SNAPSHOT",
    /**
     * The integration test failed
     */
    TEST_FAILED = "TEST_FAILED",
    /**
     * There was an error running the integration test
     */
    TEST_ERROR = "TEST_ERROR",
    /**
     * The snapshot test failed because the actual
     * snapshot was different than the expected snapshot
     */
    SNAPSHOT_FAILED = "SNAPSHOT_FAILED",
    /**
     * The snapshot test failed because there was an error executing it
     */
    SNAPSHOT_ERROR = "SNAPSHOT_ERROR",
    /**
     * The snapshot test succeeded
     */
    SNAPSHOT_SUCCESS = "SNAPSHOT_SUCCESS",
    /**
     * The integration test succeeded
     */
    TEST_SUCCESS = "TEST_SUCCESS",
    /**
     * The assertion failed
     */
    ASSERTION_FAILED = "ASSERTION_FAILED"
}